+++ src/builder.rs
@@ -6,12 +6,13 @@ use crate::error::{Error, OperatorResult};
 use crate::labels;
 use chrono::Utc;
 use k8s_openapi::api::core::v1::{
-    Affinity, ConfigMap, ConfigMapVolumeSource, Container, ContainerPort, DownwardAPIVolumeSource,
-    EmptyDirVolumeSource, EnvVar, Event, EventSource, HostPathVolumeSource, Node, NodeAffinity,
-    NodeSelector, NodeSelectorRequirement, NodeSelectorTerm, ObjectReference,
-    PersistentVolumeClaimVolumeSource, Pod, PodAffinity, PodCondition, PodSecurityContext, PodSpec,
-    PodStatus, PodTemplateSpec, Probe, ProjectedVolumeSource, ResourceRequirements, SELinuxOptions,
-    SeccompProfile, SecretVolumeSource, Sysctl, Toleration, Volume, VolumeMount,
+    Affinity, Capabilities, ConfigMap, ConfigMapVolumeSource, Container, ContainerPort,
+    DownwardAPIVolumeSource, EmptyDirVolumeSource, EnvVar, EnvVarSource, Event, EventSource,
+    HostPathVolumeSource, Node, NodeAffinity, NodeSelector, NodeSelectorRequirement,
+    NodeSelectorTerm, ObjectFieldSelector, ObjectReference, PersistentVolumeClaimVolumeSource, Pod,
+    PodAffinity, PodCondition, PodSecurityContext, PodSpec, PodStatus, PodTemplateSpec, Probe,
+    ProjectedVolumeSource, ResourceRequirements, SELinuxOptions, SeccompProfile,
+    SecretVolumeSource, SecurityContext, Sysctl, Toleration, Volume, VolumeMount,
     WindowsSecurityContextOptions,
 };
 use k8s_openapi::apimachinery::pkg::api::resource::Quantity;
@@ -20,6 +21,7 @@ use k8s_openapi::apimachinery::pkg::apis::meta::v1::{
 };
 use kube::{Resource, ResourceExt};
 use std::collections::BTreeMap;
+use std::fmt;
 use tracing::warn;
 
 /// A builder to build [`ConfigMap`] objects.
@@ -73,6 +75,168 @@ impl ConfigMapBuilder {
     }
 }
 
+/// A builder for [`SecurityContext`] objects (not to be confused with `PodSecurityContext`).
+#[derive(Clone, Default)]
+pub struct SecurityContextBuilder {
+    security_context: SecurityContext,
+}
+
+impl SecurityContextBuilder {
+    /// Convenience function for a wide use-case.
+    pub fn run_as_root() -> SecurityContext {
+        SecurityContext {
+            run_as_user: Some(0),
+            ..SecurityContext::default()
+        }
+    }
+
+    pub fn new() -> SecurityContextBuilder {
+        SecurityContextBuilder::default()
+    }
+
+    pub fn allow_privilege_escalation(&mut self, value: bool) -> &mut Self {
+        self.security_context.allow_privilege_escalation = Some(value);
+        self
+    }
+
+    pub fn capabilities(&mut self, value: Capabilities) -> &mut Self {
+        self.security_context.capabilities = Some(value);
+        self
+    }
+
+    pub fn privileged(&mut self, value: bool) -> &mut Self {
+        self.security_context.privileged = Some(value);
+        self
+    }
+
+    pub fn proc_mount(&mut self, value: impl Into<String>) -> &mut Self {
+        self.security_context.proc_mount = Some(value.into());
+        self
+    }
+
+    pub fn read_only_root_filesystem(&mut self, value: bool) -> &mut Self {
+        self.security_context.read_only_root_filesystem = Some(value);
+        self
+    }
+
+    pub fn run_as_group(&mut self, value: i64) -> &mut Self {
+        self.security_context.run_as_group = Some(value);
+        self
+    }
+
+    pub fn run_as_non_root(&mut self, value: bool) -> &mut Self {
+        self.security_context.run_as_non_root = Some(value);
+        self
+    }
+
+    pub fn run_as_user(&mut self, value: i64) -> &mut Self {
+        self.security_context.run_as_user = Some(value);
+        self
+    }
+
+    pub fn se_linux_level(&mut self, level: impl Into<String>) -> &mut Self {
+        let mut sc = self
+            .security_context
+            .se_linux_options
+            .get_or_insert_with(SELinuxOptions::default);
+        sc.level = Some(level.into());
+        self
+    }
+    pub fn se_linux_role(&mut self, role: impl Into<String>) -> &mut Self {
+        let mut sc = self
+            .security_context
+            .se_linux_options
+            .get_or_insert_with(SELinuxOptions::default);
+        sc.role = Some(role.into());
+        self
+    }
+
+    pub fn se_linux_type(&mut self, type_: impl Into<String>) -> &mut Self {
+        let mut sc = self
+            .security_context
+            .se_linux_options
+            .get_or_insert_with(SELinuxOptions::default);
+        sc.type_ = Some(type_.into());
+        self
+    }
+
+    pub fn se_linux_user(&mut self, user: impl Into<String>) -> &mut Self {
+        let mut sc = self
+            .security_context
+            .se_linux_options
+            .get_or_insert_with(SELinuxOptions::default);
+        sc.user = Some(user.into());
+        self
+    }
+
+    pub fn seccomp_profile_localhost(&mut self, profile: impl Into<String>) -> &mut Self {
+        let mut sc = self
+            .security_context
+            .seccomp_profile
+            .get_or_insert_with(SeccompProfile::default);
+        sc.localhost_profile = Some(profile.into());
+        self
+    }
+
+    pub fn seccomp_profile_type(&mut self, type_: impl Into<String>) -> &mut Self {
+        let mut sc = self
+            .security_context
+            .seccomp_profile
+            .get_or_insert_with(SeccompProfile::default);
+        sc.type_ = type_.into();
+        self
+    }
+
+    pub fn win_credential_spec(&mut self, spec: impl Into<String>) -> &mut Self {
+        let mut wo = self
+            .security_context
+            .windows_options
+            .get_or_insert_with(WindowsSecurityContextOptions::default);
+        wo.gmsa_credential_spec = Some(spec.into());
+        self
+    }
+
+    pub fn win_credential_spec_name(&mut self, name: impl Into<String>) -> &mut Self {
+        let mut wo = self
+            .security_context
+            .windows_options
+            .get_or_insert_with(WindowsSecurityContextOptions::default);
+        wo.gmsa_credential_spec_name = Some(name.into());
+        self
+    }
+
+    pub fn win_run_as_user_name(&mut self, name: impl Into<String>) -> &mut Self {
+        let mut wo = self
+            .security_context
+            .windows_options
+            .get_or_insert_with(WindowsSecurityContextOptions::default);
+        wo.run_as_user_name = Some(name.into());
+        self
+    }
+}
+
+/// Downward API capabilities available via `fieldRef`
+/// See: https://kubernetes.io/docs/tasks/inject-data-application/downward-api-volume-expose-pod-information/#capabilities-of-the-downward-api
+#[derive(Debug)]
+pub enum FieldPathEnvVar {
+    Name,
+    Namespace,
+    UID,
+    Labels(String),
+    Annotations(String),
+}
+
+impl fmt::Display for FieldPathEnvVar {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        match self {
+            FieldPathEnvVar::Name => write!(f, "metadata.name"),
+            FieldPathEnvVar::Namespace => write!(f, "metadata.namespace"),
+            FieldPathEnvVar::UID => write!(f, "metadata.uid"),
+            FieldPathEnvVar::Labels(name) => write!(f, "metadata.labels['{}']", name),
+            FieldPathEnvVar::Annotations(name) => write!(f, "metadata.annotations['{}']", name),
+        }
+    }
+}
 /// A builder to build [`Container`] objects.
 ///
 /// This will automatically create the necessary volumes and mounts for each `ConfigMap` which is added.
@@ -89,6 +253,7 @@ pub struct ContainerBuilder {
     volume_mounts: Option<Vec<VolumeMount>>,
     readiness_probe: Option<Probe>,
     liveness_probe: Option<Probe>,
+    security_context: Option<SecurityContext>,
 }
 
 impl ContainerBuilder {
@@ -113,6 +278,26 @@ impl ContainerBuilder {
         self.env.get_or_insert_with(Vec::new).push(EnvVar {
             name: name.into(),
             value: Some(value.into()),
+            ..EnvVar::default()
+        });
+        self
+    }
+
+    /// Used for pushing down attributes like the Pod's namespace into the containers.
+    pub fn add_env_var_from_field_path(
+        &mut self,
+        name: impl Into<String>,
+        field_path: FieldPathEnvVar,
+    ) -> &mut Self {
+        self.env.get_or_insert_with(Vec::new).push(EnvVar {
+            name: name.into(),
+            value_from: Some(EnvVarSource {
+                field_ref: Some(ObjectFieldSelector {
+                    field_path: field_path.to_string(),
+                    ..ObjectFieldSelector::default()
+                }),
+                ..EnvVarSource::default()
+            }),
             ..EnvVar::default()
         });
         self
@@ -191,6 +376,11 @@ impl ContainerBuilder {
         self
     }
 
+    pub fn security_context(&mut self, context: SecurityContext) -> &mut Self {
+        self.security_context = Some(context);
+        self
+    }
+
     pub fn build(&self) -> Container {
         Container {
             args: self.args.clone(),
@@ -204,6 +394,7 @@ impl ContainerBuilder {
             volume_mounts: self.volume_mounts.clone(),
             readiness_probe: self.readiness_probe.clone(),
             liveness_probe: self.liveness_probe.clone(),
+            security_context: self.security_context.clone(),
             ..Container::default()
         }
     }
@@ -927,11 +1118,17 @@ pub struct PodBuilder {
     security_context: Option<PodSecurityContext>,
     tolerations: Option<Vec<Toleration>>,
     volumes: Option<Vec<Volume>>,
+    service_account_name: Option<String>,
 }
 
 impl PodBuilder {
     pub fn new() -> PodBuilder {
         PodBuilder::default()
+    }
+
+    pub fn service_account_name(&mut self, value: impl Into<String>) -> &mut Self {
+        self.service_account_name = Some(value.into());
+        self
     }
 
     pub fn host_network(&mut self, host_network: bool) -> &mut Self {
@@ -1105,6 +1302,7 @@ impl PodBuilder {
             // In practice, this just causes a bunch of unused environment variables that may conflict with other uses,
             // such as https://github.com/stackabletech/spark-operator/pull/256.
             enable_service_links: Some(false),
+            service_account_name: self.service_account_name.clone(),
             ..PodSpec::default()
         }
     }
@@ -1357,8 +1555,8 @@ impl VolumeMountBuilder {
 mod tests {
     use crate::builder::{
         ConfigMapBuilder, ContainerBuilder, ContainerPortBuilder, EventBuilder, EventType,
-        NodeBuilder, ObjectMetaBuilder, PodBuilder, PodSecurityContextBuilder, VolumeBuilder,
-        VolumeMountBuilder,
+        FieldPathEnvVar, NodeBuilder, ObjectMetaBuilder, PodBuilder, PodSecurityContextBuilder,
+        VolumeBuilder, VolumeMountBuilder,
     };
     use k8s_openapi::api::core::v1::{
         EnvVar, Pod, PodAffinity, PodAffinityTerm, PodSecurityContext, ResourceRequirements,
@@ -1709,5 +1907,13 @@ mod tests {
             .build()
             .unwrap();
         assert_eq!(pod.metadata.name.unwrap(), "foo");
+    }
+
+    #[test]
+    pub fn test_field_ref_env_var_serialization() {
+        assert_eq!(
+            "metadata.labels['some-label-name']",
+            FieldPathEnvVar::Labels("some-label-name".to_string()).to_string()
+        );
     }
 }